webscrapping-rest/requirements.txt
fastapi
uvicorn
sqlalchemy
aiohttp
beautifulsoup4
pydantic
python-dotenv
mysql-connector-python
alembic
